add php types

+++ src/DNSMasq.php
@@ -1,4 +1,5 @@
-<?php
+<?php declare(strict_types=1);
+
 class DNSMasq {
 	private $config;
 	private $docker;
@@ -65,7 +66,8 @@ class DNSMasq {
 			$domains = [];
 
 			foreach($list as $file){
-				$contents = $this->docker->exec(containerId, "cat $file", true);
+			    $file = trim($file);
+				$contents = $this->docker->exec($containerId, "cat $file", true);
 				if(preg_match("/^[^\/]+\/(?P<domain>[^\/]+)\/(?P<ip_address>[^\/]+)/", $contents, $matches)){
 					$domains[] = ['domain' => $matches['domain'], 'ip_address' => $matches['ip_address']];
 				}
@@ -74,78 +76,96 @@ class DNSMasq {
 			return $domains;
 		}
 
-		return $this->config->getKey('dns.domains', []);
+		return $this->config->getKey('dns.domains');
 	}
 
-	public function addDomain(string $ipAddress, string $domain)
+	public function addDomain(string $ipAddress, string $domain): bool
 	{
 		$containerId = $this->docker->findRunning($this->getDockerImage());
 
-		Text::print("{blu}Installing domain:{end} '{yel}$domain{end}' with ip address '{yel}$ipAddress{end}' into dnsmasq configuration running in container '{yel}$containerId{end}'\n");
+		if(!empty($containerId)){
+            Text::print("{blu}Installing domain:{end} '{yel}$domain{end}' with ip address '{yel}$ipAddress{end}' into dnsmasq configuration running in container '{yel}$containerId{end}'\n");
 
-		$this->docker->exec($containerId, "/bin/sh -c 'echo 'address=/$domain/$ipAddress' > /etc/dnsmasq.d/$domain.conf'");
-		$this->docker->exec($containerId, "kill -s SIGHUP 1");
+            $this->docker->exec($containerId, "/bin/sh -c 'echo 'address=/$domain/$ipAddress' > /etc/dnsmasq.d/$domain.conf'");
+            $this->docker->exec($containerId, "kill -s SIGHUP 1");
 
-		sleep(2);
+            sleep(2);
 
-		$domainList = $this->config->getKey('dns.domains', []);
-		foreach($domainList as $key => $value) {
-			if($value['domain'] === $domain) unset($domainList[$key]);
-		}
-		$domainList[] = ['domain' => $domain, 'ip_address' => $ipAddress];
-		$this->config->setKey('dns.domains', array_values($domainList));
-		$this->config->write();
+            $domainList = $this->config->getKey('dns.domains');
+            foreach($domainList as $key => $value) {
+                if($value['domain'] === $domain) unset($domainList[$key]);
+            }
+            $domainList[] = ['domain' => $domain, 'ip_address' => $ipAddress];
+            $this->config->setKey('dns.domains', array_values($domainList));
+            $this->config->write();
+
+            return true;
+        }
+
+        return false;
 	}
 
 	public function removeDomain(string $domain)
 	{
 		$containerId = $this->docker->findRunning($this->getDockerImage());
 
-		Text::print("{blu}Remove domain:{end} '{yel}$domain{end}' from dnsmasq configuration running in container '{yel}$containerId{end}'\n");
+		if(!empty($containerid)){
+            Text::print("{blu}Remove domain:{end} '{yel}$domain{end}' from dnsmasq configuration running in container '{yel}$containerId{end}'\n");
 
-		$this->docker->exec($containerId, "/bin/sh -c 'f=/etc/dnsmasq.d/$domain.conf && [ -f \$f ] && rm \$f'");
-		$this->docker->exec($containerId, "kill -s SIGHUP 1");
+            $this->docker->exec($containerId, "/bin/sh -c 'f=/etc/dnsmasq.d/$domain.conf && [ -f \$f ] && rm \$f'");
+            $this->docker->exec($containerId, "kill -s SIGHUP 1");
 
-		sleep(2);
+            sleep(2);
 
-		$domainList = $this->config->getKey('dns.domains', []);
-		foreach($domainList as $key => $value) {
-			if($value['domain'] === $domain) unset($domainList[$key]);
-		}
-		$domainList = array_values($domainList);
-		$this->config->setKey('dns.domains', $domainList);
-		$this->config->write();
+            $domainList = $this->config->getKey('dns.domains');
+            foreach($domainList as $key => $value) {
+                if($value['domain'] === $domain) unset($domainList[$key]);
+            }
+            $domainList = array_values($domainList);
+            $this->config->setKey('dns.domains', $domainList);
+            $this->config->write();
+
+            return true;
+        }
+
+        return false;
 	}
 
 	public function logs(): bool
 	{
 		$container = $this->docker->findRunning($this->getDockerImage());
 
-		if($container){
+        if(!empty($containerid)){
 			$this->docker->logs($container);
 			return true;
-		}else{
-			return false;
 		}
+
+        return false;
 	}
 
 	public function logsFollow(): bool
 	{
 		$container = $this->docker->findRunning($this->getDockerImage());
 
-		if($container){
+        if(!empty($containerid)){
 			$this->docker->logsFollow($container);
 			return true;
-		}else{
-			return false;
 		}
+
+        return false;
 	}
 
+    /**
+     * @throws UnsupportedDistroException
+     */
 	public function enable(): void
 	{
 		$this->network->enableDNS();
 	}
 
+    /**
+     * @throws UnsupportedDistroException
+     */
 	public function disable(): void
 	{
 		$this->network->disableDNS();
@@ -159,6 +179,9 @@ class DNSMasq {
 		$this->docker->pull($dockerImage);
 	}
 
+    /**
+     * @throws UnsupportedDistroException
+     */
 	public function start(): void
 	{
 		if(!$this->docker->isRunning()){
@@ -181,6 +204,9 @@ class DNSMasq {
 		sleep(2);
 	}
 
+    /**
+     * @throws UnsupportedDistroException
+     */
 	public function stop(): void
 	{
 		if(!$this->docker->isRunning()){
